# blockchain-api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: KALDRIX Blockchain API
  description: REST and WebSocket API for KALDRIX blockchain operations including DAG management, quantum-resistant cryptography, and PBFT consensus
  version: 1.0.0
  contact:
    name: KALDRIX Team
    email: api@kaldrix.io
    url: https://kaldrix.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.kaldrix.io
    description: Production server
paths:
...
  /api/v1/consensus/status:
    get:
      summary: Get Consensus Status
      description: Retrieve consensus status and information
      tags:
        - Consensus
      security:
        - bearerAuth: []
      parameters:
        - name: detailed
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Consensus status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  algorithm:
                    type: string
                  status:
                    type: string
                  round:
                    type: integer
                  phase:
                    type: string
                  validators:
                    type: object
                  lastBlock:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  metrics:
                    type: object
  /api/v1/consensus/validators:
    get:
      summary: Get Validators
      description: Retrieve list of validators
      tags:
        - Consensus
      security:
        - bearerAuth: []
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Validators retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  validators:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        address:
                          type: string
                        stake:
                          type: number
                        reputation:
                          type: number
                        status:
                          type: string
                        lastActive:
                          type: string
                          format: date-time
                        blocksProduced:
                          type: integer
                        uptime:
                          type: number
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  filters:
                    type: object

  # WebSocket Events
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - bearerAuth: []
  - apiKeyAuth: []